Changed amount field to Decimal. Changed loadcontributions command. Added unit test. [#177]

+++ dc_data/dcdata/contribution/models.py
@@ -93,9 +93,7 @@ class Contribution(DataCommonsModel):
     is_amendment = models.BooleanField(default=False)
     
     # amount and date
-    amount = models.IntegerField(default=0)
-    # todo: switch to decimal once we're ready to test it
-    #amount = models.DecimalField(default=0, max_digits=15, decimal_places=2)
+    amount = models.DecimalField(default=0, max_digits=15, decimal_places=2)
     date = models.DateField(null=True)
     
     # contributor fields
